requirements.txt
# ============================================================================
# CORE - Dépendances obligatoires pour l'API
# ============================================================================

# FastAPI et serveur
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
Pillow>=10.0.0
httpx>=0.26.0

# Database (PostgreSQL for metrics)
asyncpg>=0.29.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
hypothesis>=6.92.0

# ============================================================================
# MODÈLES - Dépendances spécifiques aux modèles
# ============================================================================

# YANSNET LLM (GPT/Claude/Ollama)
openai>=1.10.0
anthropic>=0.18.0
requests>=2.31.0

# HateComment BERT - Modèle de détection de hate speech
# Dépendances spécifiques dans: app/services/hatecomment_bert/requirements.txt
# Installation: pip install -r app/services/hatecomment_bert/requirements.txt

# ML Core dependencies
numpy>=1.24.0
transformers>=4.30.0

# Autres modèles (à ajouter par les étudiants)
# Exemple pour un modèle GCN:
# torch>=2.0.0
# torch-geometric>=2.3.0
# pandas>=2.0.0
# scikit-learn>=1.3.0

# ============================================================================
# NOTES
# ============================================================================
# 
# Pour ajouter un nouveau modèle:
# 1. Créer un dossier: app/services/votre_modele/
# 2. Créer un fichier: app/services/votre_modele/requirements.txt
# 3. Ajouter vos dépendances spécifiques dans ce fichier
# 4. Installer: pip install -r app/services/votre_modele/requirements.txt
#
